FundingBandMaximum is now an int not a decimal (stored as an int so no value in passing it around as a decimal)

+++ src/Shared/SFA.DAS.SharedOuterApi/InnerApi/Responses/Apprenticeships/GetApprenticeshipPriceResponse.cs
@@ -5,7 +5,7 @@ namespace SFA.DAS.SharedOuterApi.InnerApi.Responses.Apprenticeships
     public class GetApprenticeshipPriceResponse
     {
         public Guid ApprenticeshipKey { get; set; }
-        public decimal FundingBandMaximum { get; set; }
+        public int FundingBandMaximum { get; set; }
         public decimal TrainingPrice { get; set; }
         public decimal AssessmentPrice { get; set; }
         public DateTime EarliestEffectiveDate { get; set; }
